fund-parser/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.21
anthropic>=0.40.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
